refactor to render ProjectModule(s) from map of array
the iData information for every project are now stored in an array which is run
trought a .map() which consumes the iData and return a JSX.Element that is added to
a JSX.Element[] called projectArray. this i rendered on App.jsx as <>projectArray</>
instead of having the components listed manually.

it's a minimal change that has no other purpose to show that the ProjectModule.tsx
is reusable and scalable (as now i only have to create and add a new iData obect to the array to have it rendered if i want to add a new element)